pyproject.toml
[tool.poetry]
name = "kavm-demo"
version = "0.1.0"
description = ""
authors = ["Runtime Verification Inc. <contact@runtimeverification.com>"]
readme = "README.md"
packages = [{include = "kcoin_vault"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
hypothesis = "^6.58.0"
py-algorand-sdk = "^1.20.1"
pyteal = "^0.20.1"
# beaker-pyteal = "0.4.2"
beaker-pyteal = { path = "../beaker/", develop = true }
coloredlogs = "^15.0.1"
# kavm = { git = "https://github.com/runtimeverification/avm-semantics.git", subdirectory = "kavm", branch = "integrate-autoprover-2"  }
kavm = { path = "../avm-semantics/kavm/", develop = true }

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
isort = "*"
pep8-naming = "*"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kavm-demo = 'kcoin_vault.__main__:run_demo'

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)s %(asctime)s %(name)s - %(message)s"